Migration from meta v1 to v2
